# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pirate Proof - Delivery command center and smart lock API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /lock/toggle:
    post:
      operationId: toggleLock
      tags: [lock]
      summary: Lock or unlock the smart lock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LockToggleInput"
      responses:
        "200":
          description: Updated lock status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockStatus"
...
  /lock/pins:
    get:
      operationId: listPins
      tags: [pins]
      summary: List all PIN codes
      responses:
        "200":
          description: List of PINs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pin"
    post:
      operationId: createPin
      tags: [pins]
      summary: Create a new PIN code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinInput"
      responses:
        "201":
          description: PIN created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
        "400":
          description: Bad request
  /lock/pins/{id}:
    delete:
      operationId: deletePin
      tags: [pins]
      summary: Delete a PIN code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /lock/set-pin:
    post:
      operationId: setLockPin
      tags: [lock]
      summary: Set a PIN on the IoT lock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPinInput"
      responses:
        "201":
          description: PIN created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
        "400":
          description: Bad request
...
  /email-accounts:
    get:
      operationId: listEmailAccounts
      tags: [email-accounts]
      summary: List connected email accounts
      responses:
        "200":
          description: List of email accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailAccount"
    post:
      operationId: connectEmailAccount
      tags: [email-accounts]
      summary: Connect an email account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailAccountInput"
      responses:
        "201":
          description: Email account connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAccount"
  /email-accounts/{id}:
    delete:
      operationId: deleteEmailAccount
      tags: [email-accounts]
      summary: Disconnect an email account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Disconnected
  /email-accounts/{id}/sync:
    post:
      operationId: syncEmailAccount
      tags: [email-accounts]
      summary: Scan inbox for new tracking numbers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailSyncResult"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    Delivery:
      type: object
      required: [id, userId, carrier, trackingNumber, status, deliveryModeActive, codeStatus, codeSingleUse, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        carrier:
          type: string
          enum: [amazon, ups, fedex, usps, dhl, other]
        trackingNumber:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_transit, out_for_delivery, delivered, exception, unknown]
        estimatedDelivery:
          type: ["string", "null"]
        lastLocation:
          type: ["string", "null"]
        deliveryModeActive:
          type: boolean
        accessCode:
          type: ["string", "null"]
        codeStatus:
          type: string
          enum: [none, active, used, expired]
        codeExpires:
          type: ["string", "null"]
          format: date-time
        codeSingleUse:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SetPinInput:
      type: object
      required: [label, pinCode, type]
      properties:
        label:
          type: string
        pinCode:
          type: string
          minLength: 4
          maxLength: 8
        type:
          type: string
          enum: [permanent, one_time, time_restricted, delivery]
...
    LockStatus:
      type: object
      required: [id, isLocked, lastAction, lastActionAt, deliveryModeEnabled, deliveryModeTimeout, batteryLevel]
      properties:
        id:
          type: integer
        isLocked:
          type: boolean
        lastAction:
          type: string
          enum: [locked, unlocked, pin_used, delivery_mode]
        lastActionAt:
          type: string
          format: date-time
        deliveryModeEnabled:
          type: boolean
        deliveryModeTimeout:
          type: integer
          description: Minutes before auto-relock when delivery mode triggers an unlock
        batteryLevel:
          type: integer
    LockToggleInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [lock, unlock]
...
    Pin:
      type: object
      required: [id, label, pinCode, type, isActive, usageCount, createdAt]
      properties:
        id:
          type: integer
        label:
          type: string
        pinCode:
          type: string
        type:
          type: string
          enum: [permanent, one_time, time_restricted, delivery]
        isActive:
          type: boolean
        usageCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
        validFrom:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PinInput:
      type: object
      required: [label, pinCode, type]
      properties:
        label:
          type: string
        pinCode:
          type: string
          minLength: 4
          maxLength: 8
        type:
          type: string
          enum: [permanent, one_time, time_restricted, delivery]
        expiresAt:
          type: string
        validFrom:
          type: string
        validUntil:
          type: string
...
    EmailAccount:
      type: object
      required: [id, email, provider, isActive, lastSyncAt, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        provider:
          type: string
          enum: [gmail, outlook, other]
        isActive:
          type: boolean
        lastSyncAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EmailAccountInput:
      type: object
      required: [email, provider]
      properties:
        email:
          type: string
          format: email
        provider:
          type: string
          enum: [gmail, outlook, other]
...
    EmailSyncResult:
      type: object
      required: [found, imported]
      properties:
        found:
          type: integer
        imported:
          type: integer
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/Delivery"