# lib/api-zod/src/generated/types/emailAccount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pirate Proof - Delivery command center and smart lock API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailAccountProvider } from './emailAccountProvider';

export interface EmailAccount {
  id: number;
  email: string;
  provider: EmailAccountProvider;
  isActive: boolean;
  /** @nullable */
  lastSyncAt: string | null;
  createdAt: Date;
}