# lib/api-zod/src/generated/types/emailSyncResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pirate Proof - Delivery command center and smart lock API
 * OpenAPI spec version: 0.1.0
 */
import type { Delivery } from './delivery';

export interface EmailSyncResult {
  found: number;
  imported: number;
  deliveries?: Delivery[];
}