# lib/api-zod/src/generated/types/lockStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pirate Proof - Delivery command center and smart lock API
 * OpenAPI spec version: 0.1.0
 */
import type { LockStatusLastAction } from './lockStatusLastAction';

export interface LockStatus {
  id: number;
  isLocked: boolean;
  lastAction: LockStatusLastAction;
  lastActionAt: Date;
  deliveryModeEnabled: boolean;
  /** Minutes before auto-relock when delivery mode triggers an unlock */
  deliveryModeTimeout: number;
  batteryLevel: number;
}